Move search to basic function.
Now all Chirp FS can be searched.
