+++ chirp/src/chirp_filesystem.h
@@ -120,6 +120,7 @@ INT64_T cfs_basic_swrite(int fd, const void *vbuffer, INT64_T length, INT64_T st
 INT64_T cfs_basic_putfile(const char *path, struct link * link, INT64_T length, INT64_T mode, time_t stoptime);
 INT64_T cfs_basic_getfile(const char *path, struct link * link, time_t stoptime );
 INT64_T cfs_basic_md5(const char *path, unsigned char digest[16]);
+INT64_T cfs_basic_search(const char *subject, const char *dir, const char *patt, int flags, struct link *l, time_t stoptime);
 
 /* stubs for operations not implemented in the backend FS */
 INT64_T cfs_stub_lockf (int fd, int cmd, INT64_T len);
